avito-farm-agent/requirements.txt
requests>=2.31.0
frida-tools>=12.0.0
schedule>=1.2.0
